updated the script to run submissions.

+++ scripts/set_up_inter_instance_ssh.ps1
@@ -13,7 +13,10 @@ param (
     [Parameter(Mandatory=$true)]
     [string]$instPublicAddress1,
     [Parameter(Mandatory=$true)]
-    [string]$instPrivateAddress1
+    [string]$instPrivateAddress1,
+
+    # インスタンス上でキーペアの生成を行う
+    [bool]$genKey=$true
 )
 
 $ec2UserName = "administrator"
@@ -22,26 +25,29 @@ function genAndAppendPublicKey {
     param(
         [string]$clientInstPublicAddress,
         [string]$serverInstPublicAddress,
-        [string]$serverInstPrivateAddress
+        [string]$serverInstPrivateAddress,
+        [bool]$genKey = $true
     )
 
     Write-Output "----------------------------------------------------------------"
 
-    # クライアント側インスタンスでキーペアを作成、ローカルPCに公開鍵をダウンロードする。
-    $sshCommand = "ssh-keygen -f `$home/.ssh/id_rsa -t rsa -N '`"`"'"
-    # Write-Output ${sshCommand}
-    ssh -i ${privateKey} ${ec2UserName}@${clientInstPublicAddress} ${sshCommand}
-    scp -i ${privateKey} ${ec2UserName}@${clientInstPublicAddress}:.ssh/id_rsa.pub ./id_rsa.pub
+    if ($genKey) {
+        # クライアント側インスタンスでキーペアを作成、ローカルPCに公開鍵をダウンロードする。
+        $sshCommand = "ssh-keygen -f `$home/.ssh/id_rsa -t rsa -N '`"`"'"
+        # Write-Output ${sshCommand}
+        ssh -i ${privateKey} ${ec2UserName}@${clientInstPublicAddress} ${sshCommand}
+        scp -i ${privateKey} ${ec2UserName}@${clientInstPublicAddress}:.ssh/id_rsa.pub ./id_rsa.pub
 
-    # 公開鍵の内容をサーバー側インスタンスの認証済鍵リストに追加。
-    $publicKey = Get-Content ./id_rsa.pub -encoding ascii -raw
-    Remove-Item ./id_rsa.pub
-    $sshCommand = "Add-Content `${env:ProgramData}/ssh/administrators_authorized_keys '${publicKey}'"
-    # Write-Output $sshCommand
-    ssh -i ${privateKey} ${ec2UserName}@${serverInstPublicAddress} ${sshCommand}
+        # 公開鍵の内容をサーバー側インスタンスの認証済鍵リストに追加。
+        $publicKey = Get-Content ./id_rsa.pub -encoding ascii -raw
+        Remove-Item ./id_rsa.pub
+        $sshCommand = "Add-Content `${env:ProgramData}/ssh/administrators_authorized_keys '${publicKey}'"
+        # Write-Output $sshCommand
+        ssh -i ${privateKey} ${ec2UserName}@${serverInstPublicAddress} ${sshCommand}
 
-    # SSHサーバーをリスタート。
-    ssh -i ${privateKey} ${ec2UserName}@${serverInstPublicAddress} 'Restart-Service sshd'
+        # SSHサーバーをリスタート。
+        ssh -i ${privateKey} ${ec2UserName}@${serverInstPublicAddress} 'Restart-Service sshd'
+    }
 
     # クライアント側インスタンスのknown_hostsにサーバーを登録。
     $serverPublicKeyPath = "./ssh_host_ecdsa_key.pub"
@@ -60,5 +66,5 @@ function genAndAppendPublicKey {
     ssh -i ${privateKey} ${ec2UserName}@${clientInstPublicAddress} ${sshCommand}
 }
 
-genAndAppendPublicKey ${instPublicAddress0} ${instPublicAddress1} ${instPrivateAddress1}
-genAndAppendPublicKey ${instPublicAddress1} ${instPublicAddress0} ${instPrivateAddress0}
+genAndAppendPublicKey ${instPublicAddress0} ${instPublicAddress1} ${instPrivateAddress1} ${genKey}
+genAndAppendPublicKey ${instPublicAddress1} ${instPublicAddress0} ${instPrivateAddress0} ${genKey}
